requirements.txt
aiogram>=3.13.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
